requirements.txt
numpy>=1.26,<3.0
pygame-ce>=2.5,<3.0
fpdf2>=2.7,<3.0
